backend/pyproject.toml
[project]
name = "image-catalog"
version = "0.1.0"
description = "A searchable catalog of your images"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "chromadb>=1.4.0",
    "fastapi[standard]>=0.128.0",
    "openai>=2.8.1",
    "pydantic-settings>=2.12.0",
    "pillow>=12.1.0",
    "open-clip-torch>=3.2.0",
    "python-multipart>=0.0.21",
    "numpy>=2.4.0",
    "fastmcp>=2.14.2",
]

[project.scripts]
dev = "uvicorn server:app --host 0.0.0.0 --port 8000 --reload"
index = "python indexer.py"
format = "ruff format ."
lint = "ruff check ."
lint-fix = "ruff check --fix ."
type-check = "ruff check --select=F,E,W,I"
test = "python run_tests.py"
test-unit = "python run_tests.py --unit"
test-integration = "python run_tests.py --integration"
test-cov = "python run_tests.py --coverage"
test-html = "python run_tests.py --html-cov"
test-fast = "python run_tests.py --fast"
setup = "pre-commit install"

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]
